fix: use clean code rule and split each function
